+++ site/types/schema/index.ts
@@ -1,28 +1,5 @@
-import type { Article, Career, CaseStudy, Form, Page, Report, SiteBanner, Video } from './content/index.js';
-import type { Globals, Navigation, Redirect, SEO } from './meta/index.js';
-import type { ContentType, Event, Partner, Team } from './routes/index.js';
-
-/** Directus Schema for SDK */
-export interface Schema {
-	// Routes
-	pages: Page[];
-	articles: Article[];
-	team: Team[];
-	events: Event[];
-	content_types: ContentType[];
-	partners: Partner[];
-
-	// Content
-	forms: Form[];
-	site_banners: SiteBanner[];
-	careers: Career[];
-	videos: Video[];
-	reports: Report[];
-	case_studies: CaseStudy[];
-
-	// Meta
-	navigation: Navigation[];
-	globals: Globals[];
-	redirects: Redirect[];
-	seo: SEO[];
-}
+export type * from './content/index.js';
+export type * from './meta/index.js';
+export type * from './routes/index.js';
+export type * from './schema.js';
+export type * from './system/index.js';
